// frontend/src/api/generated/models/import-node-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ImportNodeType } from './import-node-type';
import {
    ImportNodeTypeFromJSON,
    ImportNodeTypeFromJSONTyped,
    ImportNodeTypeToJSON,
    ImportNodeTypeToJSONTyped,
} from './import-node-type';

/**
 * 
 * @export
 * @interface ImportNodeJson
 */
export interface ImportNodeJson {
    /**
     * 
     * @type {string}
     * @memberof ImportNodeJson
     */
    id: string;
    /**
     * 
     * @type {ImportNodeType}
     * @memberof ImportNodeJson
     */
    type: ImportNodeType;
    /**
     * 
     * @type {string}
     * @memberof ImportNodeJson
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ImportNodeJson
     */
    url?: string;
    /**
     * 
     * @type {number}
     * @memberof ImportNodeJson
     */
    addDate?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ImportNodeJson
     */
    duplicate: boolean;
    /**
     * 
     * @type {Array<ImportNodeJson>}
     * @memberof ImportNodeJson
     */
    children?: Array<ImportNodeJson>;
}



/**
 * Check if a given object implements the ImportNodeJson interface.
 */
export function instanceOfImportNodeJson(value: object): value is ImportNodeJson {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('duplicate' in value) || value['duplicate'] === undefined) return false;
    return true;
}

export function ImportNodeJsonFromJSON(json: any): ImportNodeJson {
    return ImportNodeJsonFromJSONTyped(json, false);
}

export function ImportNodeJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImportNodeJson {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'type': ImportNodeTypeFromJSON(json['type']),
        'name': json['name'],
        'url': json['url'] == null ? undefined : json['url'],
        'addDate': json['addDate'] == null ? undefined : json['addDate'],
        'duplicate': json['duplicate'],
        'children': json['children'] == null ? undefined : ((json['children'] as Array<any>).map(ImportNodeJsonFromJSON)),
    };
}

export function ImportNodeJsonToJSON(json: any): ImportNodeJson {
    return ImportNodeJsonToJSONTyped(json, false);
}

export function ImportNodeJsonToJSONTyped(value?: ImportNodeJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': ImportNodeTypeToJSON(value['type']),
        'name': value['name'],
        'url': value['url'],
        'addDate': value['addDate'],
        'duplicate': value['duplicate'],
        'children': value['children'] == null ? undefined : ((value['children'] as Array<any>).map(ImportNodeJsonToJSON)),
    };
}